move close component
